Function sc_executor_wasmtime::create_runtime

pub fn create_runtime(
    code: &[u8], 
    heap_pages: u64, 
    host_functions: Vec<&'static dyn Function>, 
    allow_missing_func_imports: bool, 
    cache_path: Option<&Path>
) -> Result<WasmtimeRuntime, WasmError>

Create a new WasmtimeRuntime given the code. This function performs translation from Wasm to machine code, which can be computationally heavy.

The cache_path designates where this executor implementation can put compiled artifacts.